% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Tables.R
\name{Create_Tables}
\alias{Create_Tables}
\title{Create_Tables}
\usage{
Create_Tables(inputtable, metadata, taxa)
}
\arguments{
\item{inputtable}{table}

\item{metadata}{table}

\item{taxa}{table}
}
\description{
This function processes dada2 output and returns tables for plots and stats
}
