% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy_Plot_Single.R
\name{Create_Taxonomy_Plot}
\alias{Create_Taxonomy_Plot}
\title{Create_Taxonomy_Plot}
\usage{
Create_Taxonomy_Plot(taxtable, t, metadata)
}
\arguments{
\item{taxtable}{table}

\item{t}{integer}

\item{metadata}{table}
}
\description{
This function creates plots that summarize prevalent taxa
}
\details{
table of metadata.
}
