% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Stats.R
\name{diversity}
\alias{diversity}
\title{Diversity Stats}
\usage{
diversity(newmap, newtable)
}
\arguments{
\item{newmap}{table}

\item{newtable}{table}
}
\description{
This function runs stats for alpha and beta diversity
}
