% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Plots.R
\name{Diversity_Plots}
\alias{Diversity_Plots}
\title{Diversity Plots}
\usage{
Diversity_Plots(brayWmeta, newmap)
}
\arguments{
\item{brayWmeta}{table from diversity.R}

\item{newmap}{table from Create_Tables.R}
}
\description{
This function creates plots that summarize alpha and beta diversity
}
