% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Taxa_Tables.R
\name{Make_Taxa_Tables}
\alias{Make_Taxa_Tables}
\title{Make Taxa Tables
This function creates taxa tables from dada2 output
table of metadata processed by Make_Tables.}
\usage{
Make_Taxa_Tables(combined_taxa)
}
\arguments{
\item{combined_taxa}{combined table of seqtab and taxa}
}
\description{
Make Taxa Tables
This function creates taxa tables from dada2 output
table of metadata processed by Make_Tables.
}
